[FIX] 피드백 수정
